create new rabbitmq_rpc module (pylint warned that rabbitmq_utils was >1000 lines)
