+++ python/idsse_common/idsse/common/rabbitmq_utils.py
@@ -10,7 +10,6 @@
 #     Mackenzie Grimes (2)
 #
 # ----------------------------------------------------------------------------------
-
 import contextvars
 import logging
 import logging.config
@@ -278,7 +277,7 @@ class Publisher(Thread):
                   publisher is configured to confirm delivery will return False if
                   failed to confirm.
         """
-        return _blocking_publish(self.channel,
+        return blocking_publish(self.channel,
                                  self._exch,
                                  RabbitMqMessage(message, properties, route_key),
                                  self._queue)
@@ -297,6 +296,8 @@ class Publisher(Thread):
 
 class Rpc:
     """
+    !! DEPRECATED !! Use `rabbitmq_rpc.RpcPublisher` instead.
+
     RabbitMQ RPC (remote procedure call) client, runs in own thread to not block heartbeat.
     The start() and stop() methods should be called from the same thread that created the instance.
 
@@ -309,12 +310,11 @@ class Rpc:
     is not yet supported by Rpc.
 
     Example usage:
-
+        ```
         my_client = RpcClient(...insert params here...)
-
         response = my_client.send_message('{"some": "json"}')  # blocks while waiting for response
-
         logger.info(f'Response from external service: {response}')
+        ```
     """
     def __init__(self, conn_params: Conn, exch: Exch, timeout: float | None = None):
         """
@@ -333,11 +333,10 @@ class Rpc:
         self._pending_requests: dict[str, Future] = {}
 
         # Start long-running thread to consume any messages from response queue
-        self._consumer = Consumer(
-            conn_params,
-            RabbitMqParamsAndCallback(RabbitMqParams(Exch('', 'direct'), self._queue),
-                                      self._on_response)
-        )
+        self._consumer = Consumer(conn_params,
+                                  RabbitMqParamsAndCallback(
+                                      RabbitMqParams(Exch('', 'direct'), self._queue),
+                                      self._on_response))
 
     @property
     def is_open(self) -> bool:
@@ -369,7 +368,7 @@ class Rpc:
         self._pending_requests[request_id] = request_future
 
         logger.debug('Publishing request message to external service with body: %s', request_body)
-        _blocking_publish(self._consumer.channel,
+        blocking_publish(self._consumer.channel,
                           self._exch,
                           RabbitMqMessage(request_body, properties, self._exch.route_key),
                           self._queue)
@@ -437,223 +436,10 @@ class Rpc:
         return request_future.set_result(RabbitMqMessage(str(body, encoding='utf-8'), properties))
 
 
-# Tech debt: this is a temporary class "alias" to Rpc to match naming convention of
-# `RpcConsumer`. Will be deleted (and `Rpc` renamed) after current usages of `Rpc` are migrated
-class RpcPublisher(Rpc):
-    """RabbitMQ RPC (remote procedure call) publishing client, runs in own thread to not block
-    heartbeat. This class can be used to send "requests" (outbound messages) over RabbitMQ and
-    block until a "response" (inbound message) comes back from an `RpcConsumer` instance.
-    All producing to/consuming of different queues, and associating requests with their responses,
-    is abstracted away.
-
-    Note that RPC by RabbitMQ convention uses the built-in Direct Reply-To queue to field the
-    responses messages, which generates a temporary, random queue name for that individual message,
-    rather than creating its own durable queue. Directing responses to a custom, durable queue
-    is not yet supported by `RpcPublisher`.
-
-    The `start()` and `stop()` methods should be called from the same thread that created the
-    `RpcPublisher` instance.
-
-    Example usage:
-        ```
-        my_client = RpcPublisher(...insert RabbitMQ parameters here...)
-        my_client.start()
-
-        response = my_client.send_message('{"some": "json"}')  # blocks while waiting for response
-        logger.info(f'Got response from external service: {response}')
-        ```
-    """
-    # pylint: disable=arguments-renamed
-    def send_request(self, request: RabbitMqMessage) -> RabbitMqMessage | None:
-        """Send message to remote RabbitMQ service using thread-safe RPC. Will block until response
-        is received back, or timeout occurs.
-
-        Args:
-            request (RabbitMqMessage): the RabbitMQ message body and (optional) properties to send
-                as a "request" to the listening RpcConsumer service.
-
-        Returns:
-            RabbitMqMessage | None: The response message (body and properties), or None on request
-                timeout or error handling response.
-        """
-        if not self.is_open:
-            logger.debug('RPC thread not yet initialized. Setting up now')
-            self.start()
-
-        # generate unique ID to associate our request to external service's response
-        request_id = str(uuid.uuid4())
-
-        # send request to external RMQ service, providing unique RPC message ID and
-        # the queue where it should respond
-        if request.properties.headers is None:
-            request.properties.headers = {}
-        request.properties.headers['rpc'] = request_id
-        request.properties.reply_to = self._queue.name
-
-        # overwrite routing key (if any) to enforce use of default Exchange and Direct Reply-to
-        request = RabbitMqMessage(request.body, request.properties, self._exch.route_key)
-
-        # add future to dict where callback can retrieve it and set result
-        request_future = Future()
-        self._pending_requests[request_id] = request_future
-
-        logger.debug('Publishing request message to external service with body: %s', request.body)
-        _blocking_publish(self._consumer.channel,
-                          self._exch,
-                          request,
-                          self._queue)
-
-        try:
-            # block until callback runs (we'll know when the future's result has been changed)
-            return request_future.result(timeout=self._timeout)
-        except TimeoutError:
-            logger.warning('Timed out waiting for response. rpc request_id: %s', request_id)
-            self._pending_requests.pop(request_id)  # stop tracking request Future
-            return None
-        except Exception as exc:  # pylint: disable=broad-exception-caught
-            logger.warning('Unexpected response from external service: %s', str(exc))
-            self._pending_requests.pop(request_id)  # stop tracking request Future
-            return None
-
-
-class RpcResponse(NamedTuple):
-    """Data class to specify how result of RPC request should be communicated to the RMQ broker.
-    Either ack or nack with no requeue (usually a response RabbitMqMessage should be published),
-    or nack with requeue True (to re-attempt processing).
-
-    Message can be None (and is None by default), meaning request is only acked/nacked without
-    a response to the awaiting requestor, but this should generally only be used if requeue=True.
-    """
-    message: RabbitMqMessage | None
-    ack: bool = True
-    requeue: bool = False
-
-
-class RpcConsumer():
-    """Consumer RPC (remote prodecure call) class that serves as the listener to `RpcPublisher`
-    messages. `RpcConsumer` creates a thread to constantly consume RPC message "requests" emitted
-    by `RpcPublisher`, form a response, and send back it to the `RpcPublisher` asynchronously.
-
-    Note that RPC by RabbitMQ convention uses built-in Direct Reply-to queue over the default
-    exchange, and listeners for responses on a temporary queue unique to a given RPC request.
-    Publishing RpcConsumer responses to a custom, durable queue is not yet supported.
-
-    Example usage:
-        ```
-        def on_receive_request(message: RabbitMqMessage):
-            logger.info('Got request from external service: %s', message.body)
-            if message.properties.content_type == 'application/json':
-                return RpcResponse(RabbitMqMessage('success!'), ack=True)
-            return RpcResponse(None, ack=False, requeue=True)
-
-        my_consumer = RpcConsumer(<insert Conn>,
-                                  RmqParams(<insert Exch>, <insert Queue>),
-                                  on_receive_request)
-        my_consumer.start()
-        ```
-    """
-    def __init__(self,
-                 conn_params: Conn,
-                 rmq_params: RabbitMqParams,
-                 on_request_callback: Callable[[RabbitMqMessage], RpcResponse],
-                 *args,
-                 **kwargs):
-        """
-        Args:
-            conn_params (Conn): parameters to connect to RabbitMQ server
-            rmq_params (RabbitMqParams): parameters of RMQ Exchange and Queue where RPC messages
-                are expected to be received from an `RpcPublisher`.
-            on_message_callback (Callable[[RabbitMqMesssage], RpcResponse]): a function that
-                receives an inbound RPC request message, does some work with it, then returns a
-                RpcResponse, which controls if message should be acked/nacked and some
-                RabbitMqMessage published back the original requester, or if the request
-                should be nack'd and requeued to re-attempt processing
-        """
-        self._rmq_params = rmq_params
-        self._on_request_callback = on_request_callback
-
-        # Start long-running thread to consume any messages from response queue
-        self._consumer = Consumer(conn_params,
-                                  RabbitMqParamsAndCallback(rmq_params, self._on_message),
-                                  *args,
-                                  **kwargs)
-
-    @property
-    def is_open(self):
-        """Returns True if RabbitMQ connection (Consumer) is open and ready to receive messages"""
-        return self._consumer.is_alive() and self._consumer.channel.is_open
-
-    def start(self):
-        """Start dedicated threads to asynchronously receive, and send, RPC messages using a new
-        RabbitMQ connection and channel. Note: this method can be called externally, but it is
-        not required to use the client. It will automatically call this internally as needed."""
-        if not self.is_open:
-            logger.debug('Starting RPC thread to consume messages')
-            self._consumer.start()
-            self._consumer.join()
-
-    def stop(self):
-        """Unsubscribe to queue and cleanup thread(s)"""
-        logger.debug('Shutting down RpcConsumer threads')
-        if not self.is_open:
-            logger.debug('RpcConsumer threads not running, nothing to cleanup')
-            return
-
-        # tell Consumer to cleanup RabbitMQ resources and wait for thread to terminate
-        self._consumer.stop()
-        self._consumer.join()
-
-    def _on_message(self,
-                    channel: Channel,
-                    method: Basic.Deliver,
-                    properties: BasicProperties,
-                    body: bytes):
-        """Handle receiving a request message from an `RpcPublisher`. Invoke user-provided callback
-        to form response body, then send RabbitMQ message over Exchange (likely default) and Queue
-        (likely a unique Direct Reply-to) that `RpcPublisher` specified in message props.
-        """
-        request = body.decode()
-        logger.debug('Received request message from external message with body: %s', request)
-        response = self._on_request_callback(RabbitMqMessage(request, properties,
-                                                             method.routing_key))
-
-        if response.ack:
-            threadsafe_ack(channel,
-                           method.delivery_tag,
-                           lambda: logger.debug('Request %s was acked',
-                                                properties.headers.get('rpc')))
-        else:
-            threadsafe_nack(channel,
-                            method.delivery_tag,
-                            lambda: logger.debug('Request %s was nacked',
-                                                 properties.headers.get('rpc')),
-                            requeue=response.requeue)
-
-        if content := response.message:
-            # per RabbitMQ RPC convention, always use default Exchange so clients can operate on
-            # any exchange and RabbitMQ will route the message to their queue correctly
-            exch = Exch('', 'direct', route_key=properties.reply_to)
-            logger.info('Publishing response to default exchange, routing key: %s', exch.route_key)
-            logger.debug('Publishing response to external request service: %s', content.body)
-
-            # tag this response message using the "rpc" UUID from request's headers; required so
-            # RpcPublisher can associate each request/response pair and resolve all pending Futures
-            response_props = deepcopy(content.properties)
-            response_props.reply_to = properties.reply_to
-            if response_props.headers is None:
-                response_props.headers = {}
-            response_props.headers['rpc'] = properties.headers.get('rpc')
-
-            _blocking_publish(channel,
-                              exch,
-                              RabbitMqMessage(content.body, response_props, exch.route_key))
-
-
 def subscribe_to_queue(
     connection: Conn | BlockingConnection,
     rmq_params: RabbitMqParams,
-    on_message_callback: Callable[
-        [Channel, Basic.Deliver, BasicProperties, bytes], None],
+    on_message_callback: Callable[[Channel, Basic.Deliver, BasicProperties, bytes], None],
     channel: Channel | None = None
 ) -> tuple[BlockingConnection, BlockingChannel]:
     """
@@ -787,6 +573,40 @@ def threadsafe_nack(
                         extra_func)
     else:
         threadsafe_call(channel, lambda: channel.basic_nack(delivery_tag, requeue=requeue))
+
+
+def blocking_publish(
+        channel: BlockingChannel,
+        exch: Exch,
+        message_params: RabbitMqMessage,
+        queue: Queue | None = None,
+) -> bool:
+    """
+    Threadsafe, blocking publish on the specified RabbitMQ exch via the provided channel.
+    Is thread-safe.
+
+    Args:
+        channel (BlockingChannel): the pika channel to use to publish.
+        exch (Exch): parameters for the RabbitMQ exchange to publish message to.
+        message_params (RabbitMqMessage): the message body to publish, plus properties and
+        queue (optional, Queue | None): parameters for RabbitMQ queue, if message is being
+            published to a "temporary"/"private" message queue. The published message will be
+            purged from this queue after its TTL expires.
+            Default is None (destination queue not private).
+    Returns:
+        (bool) True if message published successfully. If the provided queue is confirmed to
+            confirm delivery, will return False if failed to confirm.
+    """
+    success_flag = [False]
+    done_event = Event()
+    threadsafe_call(channel, lambda: _publish(channel,
+                                              exch,
+                                              message_params,
+                                              queue,
+                                              success_flag,
+                                              done_event))
+    done_event.wait()
+    return success_flag[0]
 
 
 def _initialize_exchange_and_queue(channel: Channel, params: RabbitMqParams) -> str:
@@ -960,40 +780,6 @@ def _publish(channel: BlockingChannel,
     finally:
         if done_event:
             done_event.set()
-
-
-def _blocking_publish(
-        channel: BlockingChannel,
-        exch: Exch,
-        message_params: RabbitMqMessage,
-        queue: Queue | None = None,
-) -> bool:
-    """
-    Threadsafe, blocking publish on the specified RabbitMQ exch via the provided channel.
-    Is thread-safe.
-
-    Args:
-        channel (BlockingChannel): the pika channel to use to publish.
-        exch (Exch): parameters for the RabbitMQ exchange to publish message to.
-        message_params (RabbitMqMessage): the message body to publish, plus properties and
-        queue (optional, Queue | None): parameters for RabbitMQ queue, if message is being
-            published to a "temporary"/"private" message queue. The published message will be
-            purged from this queue after its TTL expires.
-            Default is None (destination queue not private).
-    Returns:
-        (bool) True if message published successfully. If the provided queue is confirmed to
-            confirm delivery, will return False if failed to confirm.
-    """
-    success_flag = [False]
-    done_event = Event()
-    threadsafe_call(channel, lambda: _publish(channel,
-                                              exch,
-                                              message_params,
-                                              queue,
-                                              success_flag,
-                                              done_event))
-    done_event.wait()
-    return success_flag[0]
 
 
 def _set_context(context):
